Implementação da deleção em cascata (account -> allCharacters) - funcionando

+++ lib/domain/usecases/account_usecases_impl.dart
@@ -1,4 +1,8 @@
 import 'package:injustice_app/authentication/data/repositories/i_auth_repository.dart';
+import 'package:injustice_app/core/failure/failure.dart';
+import 'package:injustice_app/core/patterns/result.dart';
+import 'package:injustice_app/data/repositories/character_repository_interface.dart';
+import 'package:injustice_app/domain/models/account_entity.dart';
 
 import '../../core/typedefs/types_defs.dart';
 import '../../data/repositories/account_repository_interface.dart';
@@ -11,8 +15,11 @@ final class GetAllAccountsUseCaseImpl implements IGetAllAccountsUseCase {
   final IAccountRepository _repository;
   final IAuthRepository _authRepository;
 
-  GetAllAccountsUseCaseImpl({required IAccountRepository repository, required IAuthRepository authRepository})
-    : _authRepository = authRepository, _repository = repository;
+  GetAllAccountsUseCaseImpl({
+    required IAccountRepository repository,
+    required IAuthRepository authRepository,
+  }) : _authRepository = authRepository,
+       _repository = repository;
 
   @override
   Future<ListAccountResult> call(NoParams params) async {
@@ -38,18 +45,20 @@ final class SaveAccountUseCaseImpl implements ISaveAccountUseCase {
   final IAccountRepository _repository;
   final IAuthRepository _authRepository;
 
-  SaveAccountUseCaseImpl({required IAccountRepository repository, required IAuthRepository authRepository})
-    : _authRepository = authRepository, _repository = repository;
+  SaveAccountUseCaseImpl({
+    required IAccountRepository repository,
+    required IAuthRepository authRepository,
+  }) : _authRepository = authRepository,
+       _repository = repository;
 
   @override
   Future<AccountResult> call(AccountParams params) async {
-
     await Future.delayed(
       const Duration(seconds: 3),
     ); // Simula um atraso para teste de loading
 
     final accountWithId = params.account.copyWith(
-      userId: _authRepository.currentSession!.user.id
+      userId: _authRepository.currentSession!.user.id,
     );
 
     return _repository.saveAccount(accountWithId);
@@ -59,13 +68,28 @@ final class SaveAccountUseCaseImpl implements ISaveAccountUseCase {
 /// usecase para deletar a conta do usuario
 final class DeleteAccountUseCaseImpl implements IDeleteAccountUseCase {
   final IAccountRepository _repository;
+  final ICharacterRepository _characterRepository;
 
-  DeleteAccountUseCaseImpl({required IAccountRepository repository})
-    : _repository = repository;
+  DeleteAccountUseCaseImpl({
+    required IAccountRepository repository,
+    required ICharacterRepository characterRepository,
+  }) : _characterRepository = characterRepository,
+       _repository = repository;
 
   @override
   Future<AccountResult> call(AccountIdParams params) async {
-    await Future.delayed(const Duration(seconds: 3)); // Simul  a um atraso para teste de loading
+    await Future.delayed(
+      const Duration(seconds: 3),
+    ); // Simula um atraso para teste de loading
+
+    final deleteCharacters = await _characterRepository.deleteAllCharacters(
+      params.id,
+    );
+
+    if (deleteCharacters.isFailure) {
+      return Error(ApiLocalFailure('Erro ao deletar personagens.'));
+    }
+
     return _repository.deleteAccount(params.id);
   }
 }
